requirements.txt
playwright>=1.45,<2.0
python-dotenv>=1.0,<2.0
loguru>=0.7,<1.0
google-genai